--- components/Header.jsx
@@ -1,30 +1,0 @@
-import React from 'react'
-import './Header.css'
-
-function Header() {
-  return (
-    <div className="header">
-        <div className="icon-name-container">
-            <div className="icon">
-
-            </div>
-
-            <div className="title">
-
-            </div>
-        </div>
-        
-        <div className="sub-title">
-
-        </div>
-        
-        <div className="fa-links">
-
-        </div>
-
-    </div>
-  );
-}
-
-export default Header
-
