+++ src/App.js
@@ -1,9 +1,10 @@
 import './App.css';
+import Header from './components/Header';
 
 function App() {
   return (
     <div className="App">
-      {/* <Header /> */}
+      <Header />
       {/* <AboutMe /> */}
       {/* <Portfolio /> */}
     </div>
