+++ src/assets/vs-myemoji-head.png
